Add search functionality to devices in Postgres
The devices table in Postgres has been improved by adding a text search feature. This includes creating an extension and indices for the name and local ID fields. Additionally, a clause for search in the SQL query in filterOptions was implemented. Corresponding changes have also been made in the test file.